+++ R/eventPred-package.R
@@ -1,6 +1,7 @@
-#' @docType package
 #' @name eventPred-package
 #' @aliases eventPred-package
+#' @keywords internal
+#' "_PACKAGE"
 #'
 #' @title Event Prediction
 #'
@@ -145,5 +146,6 @@
 #' @importFrom erify check_bool check_class check_content check_n
 #'   check_positive
 #' @importFrom rlang .data
+#' @importFrom shiny shinyAppDir
 #'
 NULL