+++ R/getPrediction.R
@@ -555,7 +555,7 @@ getPrediction <- function(
 
 
   if (!is.null(event_prior_with_covariates)) {
-    if (!is.null(covariates_event)) {
+    if (is.null(covariates_event)) {
       stop("covariates_event must be provided for event_prior_with_covariates")
     }
 
@@ -635,8 +635,6 @@ getPrediction <- function(
   }
 
 
-
-
   if (!is.null(covariates_dropout)) {
     if (is.null(df)) {
       stop("df must be provided in the presence of covariates_dropout")
@@ -654,7 +652,7 @@ getPrediction <- function(
 
 
   if (!is.null(dropout_prior_with_covariates)) {
-    if (!is.null(covariates_dropout)) {
+    if (is.null(covariates_dropout)) {
       stop(paste("covariates_dropout must be provided for",
                  "dropout_prior_with_covariates"))
     }
@@ -734,6 +732,38 @@ getPrediction <- function(
     }
   }
 
+
+  # check input data set to ensure it has all the required columns
+  if (!is.null(df)) {
+    cols = colnames(df)
+
+    if (tolower(to_predict) == "enrollment only") {
+      req_cols = c("trialsdt", "usubjid", "randdt", "cutoffdt")
+    } else {
+      req_cols = c("trialsdt", "usubjid", "randdt", "time", "event",
+                   "dropout", "cutoffdt")
+    }
+
+    if (by_treatment) {
+      req_cols <- c(req_cols, "treatment")
+    }
+
+    if (!all(req_cols %in% cols)) {
+      stop(paste("The following columns are missing from df:",
+                 paste(req_cols[!(req_cols %in% cols)], collapse = ", ")))
+    }
+
+    if (any(is.na(df[, req_cols]))) {
+      stop(paste("The following columns of df have missing values:",
+                 paste(req_cols[sapply(df, function(x) any(is.na(x)))],
+                       collapse = ", ")))
+    }
+
+    if ("treatment" %in% cols && !("treatment_description" %in% cols)) {
+      df <- df %>%
+        mutate(treatment_description = paste("Treatment", .data$treatment))
+    }
+  }
 
 
   if (!is.null(df)) {
@@ -842,74 +872,71 @@ getPrediction <- function(
       }
 
 
-      # penalized log-likelihood function with zero event
+      # penalized log-likelihood function without covariates
       lpost_exp <- function(theta, ex, d, theta0, vtheta0) {
-        -theta*d - exp(-theta)*ex - 1/(2*vtheta0)*(theta - theta0)^2
+        theta*d - exp(theta)*ex - 1/(2*vtheta0)*(theta - theta0)^2
       }
 
-      # penalized log-likelihood function with zero or 1 event
       lpost_wei <- function(theta, time, event, theta0, vtheta0) {
         shape = exp(-theta[2])
         scale = exp(theta[1])
         sum(event*dweibull(time, shape, scale, log = TRUE) +
-              (1 - event)*pweibull(time, shape, scale,
-                                   lower.tail = FALSE, log.p = TRUE)) -
-          1/2*(theta - theta0) %*% solve(vtheta0, theta - theta0)
+              (1 - event)*pweibull(
+                time, shape, scale, lower.tail = FALSE, log.p = TRUE)) -
+          1/2*as.numeric((theta - theta0) %*% solve(vtheta0, theta - theta0))
       }
 
       lpost_llogis <- function(theta, time, event, theta0, vtheta0) {
         shape = exp(-theta[2])
         scale = exp(theta[1])
-        sum(event*dllogis(time, shape, scale, log = TRUE) +
-              (1 - event)*pllogis(time, shape, scale,
-                                  lower.tail = FALSE, log.p = TRUE)) -
-          1/2*(theta - theta0) %*% solve(vtheta0, theta - theta0)
+        sum(event*flexsurv::dllogis(time, shape, scale, log = TRUE) +
+              (1 - event)*flexsurv::pllogis(
+                time, shape, scale, lower.tail = FALSE, log.p = TRUE)) -
+          1/2*as.numeric((theta - theta0) %*% solve(vtheta0, theta - theta0))
       }
 
       lpost_lnorm <- function(theta, time, event, theta0, vtheta0) {
         meanlog = theta[1]
         sdlog = exp(theta[2])
         sum(event*dlnorm(time, meanlog, sdlog, log = TRUE) +
-              (1 - event)*plnorm(time, meanlog, sdlog,
-                                 lower.tail = FALSE, log.p = TRUE)) -
-          1/2*(theta - theta0) %*% solve(vtheta0, theta - theta0)
+              (1 - event)*plnorm(
+                time, meanlog, sdlog, lower.tail = FALSE, log.p = TRUE)) -
+          1/2*as.numeric((theta - theta0) %*% solve(vtheta0, theta - theta0))
       }
 
       lpost_pwexp <- function(theta, time, event, J, tcut, theta0, vtheta0) {
         llik_pwexp(theta, time, event, J, tcut, 0, 0) -
-          1/2*(theta - theta0) %*% solve(vtheta0, theta - theta0)
+          1/2*as.numeric((theta - theta0) %*% solve(vtheta0, theta - theta0))
       }
 
 
       # penalized log-likelihood function with covariates
       if (!is.null(covariates_event)) {
-        # penalized log-likelihood function with zero event
         lpost_exp_w_x <- function(theta, time, event, q, x,
                                   theta0, vtheta0) {
-          rate = exp(-as.numeric(x %*% theta))
+          rate = exp(as.numeric(x %*% theta))
           sum(event*log(rate) - rate*time) -
-            1/2*(theta - theta0) %*% solve(vtheta0, theta - theta0)
+            1/2*as.numeric((theta - theta0) %*% solve(vtheta0, theta - theta0))
         }
 
-        # penalized log-likelihood function with zero or 1 event
         lpost_wei_w_x <- function(theta, time, event, q, x,
                                   theta0, vtheta0) {
           shape = exp(-theta[q+2])
           scale = exp(as.numeric(x %*% theta[1:(q+1)]))
           sum(event*dweibull(time, shape, scale, log = TRUE) +
-                (1 - event)*pweibull(time, shape, scale,
-                                     lower.tail = FALSE, log.p = TRUE)) -
-            1/2*(theta - theta0) %*% solve(vtheta0, theta - theta0)
+                (1 - event)*pweibull(
+                  time, shape, scale, lower.tail = FALSE, log.p = TRUE)) -
+            1/2*as.numeric((theta - theta0) %*% solve(vtheta0, theta - theta0))
         }
 
         lpost_llogis_w_x <- function(theta, time, event, q, x,
                                      theta0, vtheta0) {
           shape = exp(-theta[q+2])
           scale = exp(as.numeric(x %*% theta[1:(q+1)]))
-          sum(event*dllogis(time, shape, scale, log = TRUE) +
-                (1 - event)*pllogis(time, shape, scale,
-                                    lower.tail = FALSE, log.p = TRUE)) -
-            1/2*(theta - theta0) %*% solve(vtheta0, theta - theta0)
+          sum(event*flexsurv::dllogis(time, shape, scale, log = TRUE) +
+                (1 - event)*flexsurv::pllogis(
+                  time, shape, scale, lower.tail = FALSE, log.p = TRUE)) -
+            1/2*as.numeric((theta - theta0) %*% solve(vtheta0, theta - theta0))
         }
 
         lpost_lnorm_w_x <- function(theta, time, event, q, x,
@@ -917,15 +944,15 @@ getPrediction <- function(
           meanlog = as.numeric(x %*% theta[1:(q+1)])
           sdlog = exp(theta[q+2])
           sum(event*dlnorm(time, meanlog, sdlog, log = TRUE) +
-                (1 - event)*plnorm(time, meanlog, sdlog,
-                                   lower.tail = FALSE, log.p = TRUE)) -
-            1/2*(theta - theta0) %*% solve(vtheta0, theta - theta0)
+                (1 - event)*plnorm(
+                  time, meanlog, sdlog, lower.tail = FALSE, log.p = TRUE)) -
+            1/2*as.numeric((theta - theta0) %*% solve(vtheta0, theta - theta0))
         }
 
         lpost_pwexp_w_x <- function(theta, time, event, J, tcut, q, x,
                                     theta0, vtheta0) {
           llik_pwexp(theta, time, event, J, tcut, q, x) -
-            1/2*(theta - theta0) %*% solve(vtheta0, theta - theta0)
+            1/2*as.numeric((theta - theta0) %*% solve(vtheta0, theta - theta0))
         }
       }
 
@@ -936,14 +963,19 @@ getPrediction <- function(
 
         m = length(event_prior)
         w = sapply(event_prior, function(sub_list) sub_list$w)
+        if (any(w <= 0)) {
+          stop("w must be positive in event_prior")
+        }
         w = w/sum(w)
+
 
         # check model consistency across treatments
         model = tolower(event_prior[[1]]$model)
         if (m > 1) {
           for (j in 2:m) {
             if (tolower(event_prior[[j]]$model) != model) {
-              stop("Prior event model must be equal across treatments")
+              stop(paste("Event model must be equal across treatments",
+                         "in event_prior"))
             }
           }
 
@@ -952,7 +984,7 @@ getPrediction <- function(
               if (!all.equal(event_prior[[j]]$piecewiseSurvivalTime,
                              event_prior[[1]]$piecewiseSurvivalTime)) {
                 stop(paste("piecewiseSurvivalTime must be equal",
-                           "across treatments"))
+                           "across treatments in event_prior"))
               }
             }
           }
@@ -973,6 +1005,7 @@ getPrediction <- function(
                                     t(event_prior[[j]]$theta))
         }
         vtheta = vtheta - theta %*% t(theta)
+        if (length(theta) == 1) vtheta = as.numeric(vtheta)
 
         if (model %in% c("exponential", "weibull", "log-logistic",
                          "log-normal")) {
@@ -987,12 +1020,14 @@ getPrediction <- function(
       }
 
 
-
       if (!is.null(event_prior_with_covariates) && !by_treatment &&
           !("model" %in% names(event_prior_with_covariates))) {
 
         m = length(event_prior_with_covariates)
         w = sapply(event_prior_with_covariates, function(sub_list) sub_list$w)
+        if (any(w <= 0)) {
+          stop("w must be positive in event_prior_with_covariates")
+        }
         w = w/sum(w)
 
         # check model consistency across treatments
@@ -1000,7 +1035,8 @@ getPrediction <- function(
         if (m > 1) {
           for (j in 2:m) {
             if (tolower(event_prior_with_covariates[[j]]$model) != model) {
-              stop("Prior event model must be equal across treatments")
+              stop(paste("Event model must be equal across treatments",
+                         "in event_prior_with_covariates"))
             }
           }
 
@@ -1010,7 +1046,7 @@ getPrediction <- function(
                 event_prior_with_covariates[[j]]$piecewiseSurvivalTime,
                 event_prior_with_covariates[[1]]$piecewiseSurvivalTime)) {
                 stop(paste("piecewiseSurvivalTime must be equal",
-                           "across treatments"))
+                           "across treatments in event_prior_with_covariates"))
               }
             }
           }
@@ -1049,10 +1085,11 @@ getPrediction <- function(
 
       # fit the event model without covariates
       if (!(to_predict == "event only" && !is.null(covariates_event))) {
+        event_fit <- fitEvent(df, event_model,
+                              piecewiseSurvivalTime, k, scale,
+                              showplot, by_treatment)
+
         if (is.null(event_prior)) {  # no prior, use MLE
-          event_fit <- fitEvent(df, event_model,
-                                piecewiseSurvivalTime, k, scale,
-                                showplot, by_treatment)
           event_fit1 <- event_fit$event_fit
         } else {
           if (!by_treatment) {
@@ -1077,8 +1114,8 @@ getPrediction <- function(
                             theta0 = event_prior2[[j]]$theta,
                             vtheta0 = event_prior2[[j]]$vtheta,
                             method = "Brent",
-                            lower = event_prior2[[j]]$theta-10,
-                            upper = event_prior2[[j]]$theta+10,
+                            lower = event_prior2[[j]]$theta - 2,
+                            upper = event_prior2[[j]]$theta + 2,
                             control = c(fnscale = -1),
                             hessian = TRUE)  # maximization
               event_fit2$model = "Exponential"
@@ -1146,14 +1183,14 @@ getPrediction <- function(
       }
 
 
-
       # fit the event model with covariates
       if (!is.null(covariates_event)) {
+        event_fit_w_x <- fitEvent(df, event_model,
+                                  piecewiseSurvivalTime, k, scale,
+                                  showplot, by_treatment,
+                                  covariates_event)
+
         if (is.null(event_prior_with_covariates)) {  # no prior, use MLE
-          event_fit_w_x <- fitEvent(df, event_model,
-                                    piecewiseSurvivalTime, k, scale,
-                                    showplot, by_treatment,
-                                    covariates_event)
           event_fit1_w_x <- event_fit_w_x$event_fit
         } else {
           if (!by_treatment) {
@@ -1187,6 +1224,7 @@ getPrediction <- function(
               opt1 <- optim(event_prior2_w_x[[j]]$theta,
                             lpost_wei_w_x, gr = NULL,
                             time = df1$time, event = df1$event,
+                            q = q_event, x = x_event,
                             theta0 = event_prior2_w_x[[j]]$theta,
                             vtheta0 = event_prior2_w_x[[j]]$vtheta,
                             control = c(fnscale = -1),
@@ -1198,6 +1236,7 @@ getPrediction <- function(
               opt1 <- optim(event_prior2_w_x[[j]]$theta,
                             lpost_llogis_w_x, gr = NULL,
                             time = df1$time, event = df1$event,
+                            q = q_event, x = x_event,
                             theta0 = event_prior2_w_x[[j]]$theta,
                             vtheta0 = event_prior2_w_x[[j]]$vtheta,
                             control = c(fnscale = -1),
@@ -1209,6 +1248,7 @@ getPrediction <- function(
               opt1 <- optim(event_prior2_w_x[[j]]$theta,
                             lpost_lnorm_w_x, gr = NULL,
                             time = df1$time, event = df1$event,
+                            q = q_event, x = x_event,
                             theta0 = event_prior2_w_x[[j]]$theta,
                             vtheta0 = event_prior2_w_x[[j]]$vtheta,
                             control = c(fnscale = -1),
@@ -1246,8 +1286,6 @@ getPrediction <- function(
       }
 
 
-
-
       # whether to include dropout model
       if (tolower(dropout_model) != "none") {
 
@@ -1257,6 +1295,9 @@ getPrediction <- function(
 
           m = length(dropout_prior)
           w = sapply(dropout_prior, function(sub_list) sub_list$w)
+          if (any(w <= 0)) {
+            stop("w must be positive in dropout_prior")
+          }
           w = w/sum(w)
 
           # check model consistency across treatments
@@ -1310,7 +1351,6 @@ getPrediction <- function(
         }
 
 
-
         if (!is.null(dropout_prior_with_covariates) && !by_treatment &&
             !("model" %in% names(dropout_prior_with_covariates))) {
 
@@ -1356,6 +1396,7 @@ getPrediction <- function(
                  t(dropout_prior_with_covariates[[j]]$theta))
           }
           vtheta = vtheta - theta %*% t(theta)
+          if (length(theta) == 1) vtheta = as.numeric(vtheta)
 
           if (model %in% c("exponential", "weibull", "log-logistic",
                            "log-normal")) {
@@ -1372,14 +1413,14 @@ getPrediction <- function(
         }
 
 
-
         # fit the dropout model without covariates
         if (!(to_predict == "event only" && !is.null(covariates_dropout))) {
+          dropout_fit <- fitDropout(df, dropout_model,
+                                    piecewiseDropoutTime,
+                                    k_dropout, scale_dropout,
+                                    showplot, by_treatment)
+
           if (is.null(dropout_prior)) {  # no prior, use MLE
-            dropout_fit <- fitDropout(df, dropout_model,
-                                      piecewiseDropoutTime,
-                                      k_dropout, scale_dropout,
-                                      showplot, by_treatment)
             dropout_fit1 <- dropout_fit$dropout_fit
           } else {
             if (!by_treatment) {
@@ -1404,8 +1445,8 @@ getPrediction <- function(
                               theta0 = dropout_prior2[[j]]$theta,
                               vtheta0 = dropout_prior2[[j]]$vtheta,
                               method = "Brent",
-                              lower = dropout_prior2[[j]]$theta-10,
-                              upper = dropout_prior2[[j]]$theta+10,
+                              lower = dropout_prior2[[j]]$theta - 2,
+                              upper = dropout_prior2[[j]]$theta + 2,
                               control = c(fnscale = -1),
                               hessian = TRUE)  # maximization
                 dropout_fit2$model = "Exponential"
@@ -1473,14 +1514,14 @@ getPrediction <- function(
         }
 
 
-
         # fit the dropout model with covariates
         if (!is.null(covariates_dropout)) {
+          dropout_fit_w_x <- fitDropout(
+            df, dropout_model, piecewiseDropoutTime,
+            k_dropout, scale_dropout, showplot, by_treatment,
+            covariates_dropout)
+
           if (is.null(dropout_prior_with_covariates)) {  # no prior, use MLE
-            dropout_fit_w_x <- fitDropout(
-              df, dropout_model, piecewiseDropoutTime,
-              k_dropout, scale_dropout, showplot, by_treatment,
-              covariates_dropout)
             dropout_fit1_w_x <- dropout_fit_w_x$dropout_fit
           } else {
             if (!by_treatment) {
@@ -1514,6 +1555,7 @@ getPrediction <- function(
                 opt1 <- optim(dropout_prior2_w_x[[j]]$theta,
                               lpost_wei_w_x, gr = NULL,
                               time = df1$time, event = df1$dropout,
+                              q = q_dropout, x = x_dropout,
                               theta0 = dropout_prior2_w_x[[j]]$theta,
                               vtheta0 = dropout_prior2_w_x[[j]]$vtheta,
                               control = c(fnscale = -1),
@@ -1525,6 +1567,7 @@ getPrediction <- function(
                 opt1 <- optim(dropout_prior2_w_x[[j]]$theta,
                               lpost_llogis_w_x, gr = NULL,
                               time = df1$time, event = df1$dropout,
+                              q = q_dropout, x = x_dropout,
                               theta0 = dropout_prior2_w_x[[j]]$theta,
                               vtheta0 = dropout_prior2_w_x[[j]]$vtheta,
                               control = c(fnscale = -1),
@@ -1536,6 +1579,7 @@ getPrediction <- function(
                 opt1 <- optim(dropout_prior2_w_x[[j]]$theta,
                               lpost_lnorm_w_x, gr = NULL,
                               time = df1$time, event = df1$dropout,
+                              q = q_dropout, x = x_dropout,
                               theta0 = dropout_prior2_w_x[[j]]$theta,
                               vtheta0 = dropout_prior2_w_x[[j]]$vtheta,
                               control = c(fnscale = -1),
@@ -1571,7 +1615,6 @@ getPrediction <- function(
         } else {
           dropout_fit1_w_x <- NULL
         }
-
 
 
         if (grepl("enrollment", to_predict, ignore.case = TRUE)) {
@@ -1860,5 +1903,3 @@ getPrediction <- function(
     }
   }
 }
-
-